+++ CHANGELOG.md
@@ -49,6 +49,7 @@
     + `-[NSMutableArray matchWith:]`
     + NSDictionary methods: `-esObjectForKey`, `-each:`, `-each:option:`, `match:`, `matchDictionary:`, `matches:`, `matches:option:`, `matchWith:`, `writeToFile:atomically:completion:`
     + NSOrderedSet methods: `-each:`, `-each:option:`, `-match:`, `matchWith:`
+    + NSSet methods: `each`, `match`, `matches`, `matchWith`
 - Refactored:
     + `NSObject (ESAutoCoding)`
     + `-queryDictionary` of `NSURL`/`NSString` to use the `NSURLComponents` API